requirements.txt
amqp==2.5.2
anyjson==0.3.3
apturl==0.5.2
asgiref==3.2.3
asn1crypto==0.24.0
Babel==2.8.0
bcrypt==3.1.7
billiard==3.6.2.0
bottle==0.12.13
Brlapi==0.6.6
celery==3.1.26.post2
certifi==2018.1.18
cffi==1.14.0
chardet==3.0.4
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==2.1.3
django-celery==3.2.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docker==2.5.1
docker-pycreds==0.2.1
flower==0.9.3
gevent==1.2.2
Glances==2.11.1
greenlet==0.4.15
httplib2==0.9.2
idna==2.6
importlib-metadata==1.5.0
influxdb==4.1.1
keyring==10.6.0
keyrings.alt==3.0
kombu==4.6.7
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.1.1
netifaces==0.10.4
numpy==1.13.3
oauth==1.0.1
olefile==0.45.1
paramiko==2.7.1
pexpect==4.2.1
Pillow==5.1.0
ply==3.11
protobuf==3.0.0
psutil==5.4.2
psycopg2==2.8.4
pyasn1==0.4.2
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycryptodomex==3.4.7
pycups==1.9.73
pygobject==3.26.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.2.0
pyRFC3339==1.0
pysmi==0.2.2
pysnmp==4.4.3
pystache==0.5.4
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.6.1
python-debian==0.1.32
python-memcached==1.53
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
redis==2.10.6
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scp==0.13.2
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.14.0
sqlparse==0.3.0
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tornado==6.0.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
vine==1.3.0
virtualenv==15.1.0
wadllib==1.3.2
websocket-client==0.44.0
xkit==0.0.0
zipp==3.0.0
zope.interface==4.3.2